style: simplified gifs display by displaying WebP for all browsers
Fallback to gifs (via a 'picture' markup) was causing Safari a weird bug.
When loading an image, its containing div wasn't getting resized appropriately.
This is probably a Safari bug, as the div is appropriately resized when rotating the screen,
or quitting re-entring safari. Nonetheless, webP has >95% browser compatibility at this time of writing.